pyproject.toml
[project]
name = "couch-async"
version = "0.1.1"
description = "Async HTTP/2 client library for CouchDB 3.x with type-safe interfaces"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Erik van Raalte", email = "erik.vanraalte94@gmail.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Typing :: Typed",
]
keywords = ["couchdb", "async", "httpx", "http2", "pydantic"]
dependencies = [
    "httpx[http2]>=0.28.1",
    "pydantic-settings>=2.12.0",
]

[project.urls]
Repository = "https://github.com/evanraalte/couch-async"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
    "ty>=0.0.16",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["couch"]

[tool.ruff.lint]
select = ['ALL']
ignore = [
    'D', # Docstrings can be omitted for now
    "T201",    # Prints are not OK, but will be done by OTEL
    "COM812", # recommended by ruff
    "TRY003", # exception messages
    "EM101", # string literal
    "S101", # assertions are okay
    "RET504",
    "ANN401", # Any types needed for CouchDB's dynamic key/value types
    "PLR0913", # view function needs many params for CouchDB API
    "C901", # view function complexity is acceptable
]

[tool.ruff.lint.extend-per-file-ignores]
"tests/*" = [
    "S106", # hardcoded values
    "PLR2004", # magic values
    "ANN201", # return type annotation
    "SLF001", # private members
    "S501", # unsafe https
    "PLR2004",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
